Split ItemMode::treatItemsAsFolders into two modes
The two modes are:
- ItemMode::treatItemsAsFolders
- ItemMode::treatItemsAsFoldersWithFileBumping

Now, ItemMode::treatItemsAsFolders will simply treat
all items as folders without looking inside them.

ItemMode::treatItemsAsFoldersWithFileBumping has the behavior
of the old ItemMode::treatItemsAsFolders. It checks the contents
of every item, and if an item contains only one file with the same
name, it bumps the file up one level and removes the item. This can
be convenient to use, but it may result in a lot of API calls and
a big slowdown if the internet connection is slow.