Removed almost every possible root path api call
Now, the root path api call will only be performed
if we change folders to another folder that is not
in the current directory and not within our root
path. This currently does not really happen.

This still needs some more testing, but it seems
to work fine. This reduces the number of api calls
to 2 in most cases.